fix: share CLI error with local tmux tests
(cherry picked from commit fc1dc8a5e7cb19c02a0e9f11c233a08cf3139c32)
